Init rework
