+++ radiomics/src/cuda/test.cu
@@ -41,7 +41,9 @@ int AddShape2DFunction(size_t idx, shape_2D_func_t func) {
 }
 
 SOLUTION_DECL(0);
+SOLUTION_DECL(1);
 
 void RegisterSolutions() {
     REGISTER_SOLUTION(0);
+    REGISTER_SOLUTION(1);
 }
